backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database (MongoDB async)
motor==3.3.2
pymongo==4.6.1

# Redis cache (async)
redis[hiredis]==5.0.1

# NLP and ML
transformers==4.37.0
sentence-transformers==2.3.1
torch==2.1.2
hdbscan==0.8.33
scikit-learn==1.4.0

# Article fetching
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
aiohttp==3.9.1
newspaper3k==0.2.8

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
slowapi==0.1.9
numpy==1.26.3
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3
pytest-cov==4.1.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
